requirements.txt
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==41.0.5
dropbox==10.10.0
google-api-core==2.20.0
google-api-python-client==2.51.0
google-auth==2.16.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.28.2
pushbullet.py==0.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.1.4
python-dotenv==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.7.0
setuptools==75.1.0
six==1.16.0
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
